# data-service/openapi/admin.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/admin/portfolio/sections": {
      "get": {
        "operationId": "SectionsController_list",
        "summary": "List the draft sections (FR-CFG-1)",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AdminSectionDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key or user id."
          }
        },
        "tags": [
          "sections"
        ],
        "security": [
          {
            "api-key": []
          }
        ]
      }
    },
    "/admin/portfolio/sections/{type}": {
      "patch": {
        "operationId": "SectionsController_update",
        "summary": "Toggle, reorder, or replace a section's content (FR-CFG-3)",
        "parameters": [
          {
            "name": "type",
            "required": true,
            "in": "path",
            "schema": {
              "enum": [
                "hero",
                "projects",
                "skills",
                "contact",
                "experience",
                "education",
                "blog",
                "testimonials",
                "opensource",
                "speaking",
                "achievements",
                "trainings",
                "gallery"
              ],
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSectionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminSectionDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key or user id."
          }
        },
        "tags": [
          "sections"
        ],
        "security": [
          {
            "api-key": []
          }
        ]
      }
    },
    "/admin/portfolio/sections/{type}/items": {
      "post": {
        "operationId": "SectionsController_addItem",
        "summary": "Add an item to a collection section (FR-API-4)",
        "parameters": [
          {
            "name": "type",
            "required": true,
            "in": "path",
            "schema": {
              "enum": [
                "hero",
                "projects",
                "skills",
                "contact",
                "experience",
                "education",
                "blog",
                "testimonials",
                "opensource",
                "speaking",
                "achievements",
                "trainings",
                "gallery"
              ],
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "One item, shaped by the registry's itemFields for :type.",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminSectionDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key or user id."
          }
        },
        "tags": [
          "sections"
        ],
        "security": [
          {
            "api-key": []
          }
        ]
      }
    },
    "/admin/portfolio/sections/{type}/items/{itemId}": {
      "patch": {
        "operationId": "SectionsController_updateItem",
        "summary": "Update an item in a collection section (FR-API-4)",
        "parameters": [
          {
            "name": "type",
            "required": true,
            "in": "path",
            "schema": {
              "enum": [
                "hero",
                "projects",
                "skills",
                "contact",
                "experience",
                "education",
                "blog",
                "testimonials",
                "opensource",
                "speaking",
                "achievements",
                "trainings",
                "gallery"
              ],
              "type": "string"
            }
          },
          {
            "name": "itemId",
            "required": true,
            "in": "path",
            "schema": {
              "example": "66e1f0c2a1b2c3d4e5f60718",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "One item, shaped by the registry's itemFields for :type.",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminSectionDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key or user id."
          }
        },
        "tags": [
          "sections"
        ],
        "security": [
          {
            "api-key": []
          }
        ]
      },
      "delete": {
        "operationId": "SectionsController_removeItem",
        "summary": "Remove an item from a collection section (FR-API-4)",
        "parameters": [
          {
            "name": "type",
            "required": true,
            "in": "path",
            "schema": {
              "enum": [
                "hero",
                "projects",
                "skills",
                "contact",
                "experience",
                "education",
                "blog",
                "testimonials",
                "opensource",
                "speaking",
                "achievements",
                "trainings",
                "gallery"
              ],
              "type": "string"
            }
          },
          {
            "name": "itemId",
            "required": true,
            "in": "path",
            "schema": {
              "example": "66e1f0c2a1b2c3d4e5f60718",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Removed."
          },
          "401": {
            "description": "Missing or invalid API key or user id."
          }
        },
        "tags": [
          "sections"
        ],
        "security": [
          {
            "api-key": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Portfolio API — admin",
    "description": "API-key authenticated. Reads and writes the draft of the caller's portfolio (SRS §7.2).",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ],
  "components": {
    "securitySchemes": {
      "api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key"
      }
    },
    "schemas": {
...
      "AdminSectionDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "hero",
              "projects",
              "skills",
              "contact",
              "experience",
              "education",
              "blog",
              "testimonials",
              "opensource",
              "speaking",
              "achievements",
              "trainings",
              "gallery"
            ],
            "example": "projects"
          },
          "enabled": {
            "type": "boolean",
            "example": true
          },
          "order": {
            "type": "number",
            "example": 1
          },
          "content": {
            "type": "object",
            "additionalProperties": true,
            "description": "Shaped by the registry's descriptor for `type`.",
            "example": {
              "items": []
            }
          }
        },
        "required": [
          "type",
          "enabled",
          "order",
          "content"
        ]
      },
...
      "UpdateSectionDto": {
        "type": "object",
        "properties": {
          "enabled": {
            "type": "boolean",
            "example": true
          },
          "order": {
            "type": "number",
            "example": 2
          },
          "content": {
            "type": "object",
            "additionalProperties": true,
            "description": "Replacement content, shaped by the registry's descriptor for :type.",
            "example": {
              "items": []
            }
          }
        }
      },
...
    }
  }
}